// ui-app/src/app/api/models/dr-sheet-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { StudentDetails } from '../models/student-details';
export interface DrSheetResponse {
  className?: string;
  examCenterName?: string;
  students?: Array<StudentDetails>;
  vigyanKendraCode?: string;
  vigyanKendraName?: string;
}